Fix few issue with sol/settings/logging/CI
