+++ .github/workflows/SegsNightlyRelease.yaml
@@ -85,14 +85,14 @@ jobs:
 
   build-linux:
     needs: tidy-up-releases
-    runs-on: 'ubuntu-latest'
+    runs-on: 'ubuntu-24.04'
     permissions:
       contents: write
     steps:
     - uses: actions/checkout@v3
 
     - name: Install Qt Linux
-      uses: jurplel/install-qt-action@v2.14.0
+      uses: jurplel/install-qt-action@v4
       with:
         host: linux
         target: desktop
@@ -103,7 +103,8 @@ jobs:
     - name: Configure CMake Linux
       env:
         CMAKE_PREFIX_PATH: ${{ env.Qt6_Dir }}
-        CXX: "/usr/bin/clang++"
+        CXX: "/usr/bin/g++-13"
+        CC: "/usr/bin/gcc-13"
       run: cmake -B ${{ github.workspace }}/build -DCMAKE_BUILD_TYPE=${{ env.BUILD_TYPE }}
 
     - name: Build Linux
