+++ Data/settings_template.cfg
@@ -27,18 +27,20 @@
 config_version     = 2
 
 [AdminServer]
-AccountDatabase\db_driver   = QSQLITE
-AccountDatabase\db_host     = 127.0.0.1
-AccountDatabase\db_port     = 5432
-AccountDatabase\db_name     = segs.db
-AccountDatabase\db_user     = segsadmin
-AccountDatabase\db_pass     = segs123
-CharacterDatabase\db_driver = QSQLITE
-CharacterDatabase\db_host   = 127.0.0.1
-CharacterDatabase\db_port   = 5432
-CharacterDatabase\db_name   = segs_game.db
-CharacterDatabase\db_user   = segsadmin
-CharacterDatabase\db_pass   = segs123
+[AdminServer/AccountDatabase]
+db_driver   = QSQLITE
+db_host     = 127.0.0.1
+db_port     = 5432
+db_name     = segs.db
+db_user     = segsadmin
+db_pass     = segs123
+[AdminServer/CharacterDatabase]
+db_driver = QSQLITE
+db_host   = 127.0.0.1
+db_port   = 5432
+db_name   = segs_game.db
+db_user   = segsadmin
+db_pass   = segs123
 
 [AuthServer]
 location_addr           = 127.0.0.1:2106
