+++ Servers/AuthServer/CMakeLists.txt
@@ -25,7 +25,6 @@ SET (target_SOURCES
   ${target_INCLUDE}
 )
 add_handler(Auth ${target_SOURCES})
-#FIND_PACKAGE(Boost COMPONENTS system REQUIRED) # Boost_LIBRARIES is not a cache variable, so we  have to do this here again
 
 target_link_libraries(Auth PUBLIC Qt6::Core)
 target_link_libraries(Auth PUBLIC ace_IMP cox_server_common AuthProtocol SEGS_Components)
@@ -47,7 +46,7 @@ IF(BUILD_SEPARATE)
     )
     ADD_DEPENDENCIES(segs_server ${target_DEPENDS} Auth)
     ADD_DEPENDENCIES(segs_server ACE_BUILD)
-    target_link_libraries(segs_server PUBLIC Qt6::Core)
+    target_link_libraries(segs_server PUBLIC Qt6::Core SEGS_Env)
     set_property(TARGET segs_server PROPERTY POSITION_INDEPENDENT_CODE ON)
     IF(MINGW)
         target_link_libraries(segs_server PUBLIC ws2_32)
